backend/requirements.txt
fastapi
uvicorn[standard]
requests
pandas
ta
groq
python-dotenv
